RuleSetBaseを整理
