+++ VitDeck/Assets/VitDeck/Validator/Rules/Vket4/DebugTargetEnumerationRule.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 080a790812c18a143ab6afab73eced33
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
